requirements.txt
anyio==3.7.0
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
dateparser==1.1.8
exceptiongroup==1.1.1
GoogleNews==1.6.8
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tornado==6.3.2
tzlocal==5.0.1
urllib3==2.0.2
